requirements.txt
discord.py>=2.3.1
APScheduler>=3.10.1
python-dotenv>=1.0.0
